+++ libraries/compute-utils/function-views/src/shared/utils.ts
@@ -29,12 +29,12 @@ export function getObservable<T>(onInput: (f: Function) => SubscriptionLike): Ob
 export const deepCopy = (call: DG.FuncCall) => {
   const deepClone = call.clone();
 
-  const dfOutputs = wu(call.outputParams.values() as DG.FuncCallParam[])
+  const dfOutputs = wu(call.outputParams.values())
     .filter((output) => output.property.propertyType === DG.TYPE.DATA_FRAME);
   for (const output of dfOutputs)
     deepClone.outputs[output.name] = call.outputs[output.name].clone();
 
-  const dfInputs = wu(call.inputParams.values() as DG.FuncCallParam[])
+  const dfInputs = wu(call.inputParams.values())
     .filter((input) => input.property.propertyType === DG.TYPE.DATA_FRAME);
   for (const input of dfInputs)
     deepClone.inputs[input.name] = call.inputs[input.name].clone();