+++ packages/Bio/src/utils/poly-tool/ui.ts
@@ -10,6 +10,8 @@ import {addTransformedColumn} from './transformation';
 import * as rxjs from 'rxjs';
 //import {MetaData} from './types';
 
+const RULE_PATH = 'System:AppData/Bio/polytool-rules/';
+
 export function getPolyToolDialog(): DG.Dialog {
   //const monomerLib = MonomerLibManager.instance.getBioLib();
   const targetColumns = grok.shell.t.columns.bySemTypeAll(DG.SEMTYPE.MACROMOLECULE);
@@ -24,41 +26,19 @@ export function getPolyToolDialog(): DG.Dialog {
   const generateHelmChoiceInput = ui.boolInput('Get HELM', true);
   ui.tooltip.bind(generateHelmChoiceInput.root, 'Add HELM column');
 
-  let rulesTable: DG.DataFrame = DG.DataFrame.create();
+  // let rulesTable: DG.DataFrame = DG.DataFrame.create();
 
   const ruleFileInput = ui.button('ADD RULES', () => {
     DG.Utils.openFile({
       accept: '.csv',
       open: async (selectedFile) => {
         const content = await selectedFile.text();
-        rulesTable = DG.DataFrame.fromCsv(content);
+        // rulesTable = DG.DataFrame.fromCsv(content);
+        await grok.dapi.files.writeAsText(RULE_PATH + `${selectedFile.name}`, content);
         //console.log(df.toCsv());
       },
     });
   });
-  // dialog.addButton(
-  //   'Add',
-  //   () => eventManager.addLibraryFile(),
-  //   undefined,
-  //   'Upload new HELM monomer library'
-  // );
-
-
-  // const ruleFileInput = DG.Utils.openFile({
-  //   accept: '.csv',
-  //   open: async (selectedFile) => {
-  //     const content = await selectedFile.text();
-  //     const name = selectedFile.name;
-  //     const df = DG.DataFrame.fromCsv(content);
-
-  //     console.log(df.toCsv());
-  //   },
-  // });
-
-  //grok.data.files.openTable('Samples:Files/chem/smiles_10K_with_activities.csv')
-  // const file = await loadFileAsText(tableName);
-  // const df = DG.DataFrame.fromCsv(file);
-  // df.name = tableName.replace('.csv', '');
 
   const div = ui.div([
     targetColumnInput,
@@ -74,7 +54,7 @@ export function getPolyToolDialog(): DG.Dialog {
         grok.shell.warning('No marcomolecule column chosen!');
         return;
       }
-      addTransformedColumn(molCol!, rulesTable!, generateHelmChoiceInput.value!);
+      addTransformedColumn(molCol!, generateHelmChoiceInput.value!);
     }
     );