+++ packages/CVMTests/src/scripts/scripts_tests.ts
@@ -52,12 +52,15 @@ for (const lang of langs) {
     });
 
     test('Dataframe input/output', async () => {
-      const df = DataFrame.fromCsv(`id,date,name\nid1,${Date.now()},datagrok`);
+      function getSample() {
+        return DataFrame.fromCsv(`id,date,name\nid1,${Date.now()},datagrok`)
+      }
       const result = await grok.functions.call(`CVMTests:${lang}Dataframe`,
-        {'df': df, 'dfNumerical': df, 'dfCategorical': df});
-      expectTable(result['resultDf'], df);
-      expectTable(result['resultNumerical'], df);
-      expectTable(result['resultCategorical'], df);
+        {'df': getSample(), 'dfNumerical': getSample(), 'dfCategorical': getSample()});
+      const sample = getSample();
+      expectTable(result['resultDf'], sample);
+      expectTable(result['resultNumerical'], sample);
+      expectTable(result['resultCategorical'], sample);
     });
 
     test('Map type input/output', async () => {
@@ -145,7 +148,7 @@ for (const lang of langs) {
       const start = Date.now();
       await df.columns.addNewCalculated('new', `CVMTests:${lang}CalcColumn(\${age})`);
       return `Execution time: ${Date.now() - start}`;
-    }, {timeout: 120000, skipReason: ['Octave', 'NodeJS', 'Julia'].includes(lang) ? 'GROK-13876' : undefined});
+    }, {timeout: 120000});
 
     test(`Dataframe performance test sequentially`, async () => {
       const iterations = DG.Test.isInBenchmark ? 10 : 3;
@@ -157,7 +160,7 @@ for (const lang of langs) {
       const sum = results.reduce((p, c) => p + c, 0);
       return toDart({'Average time': sum / results.length,
         'Min time': Math.min(...results), 'Max time': Math.max(...results)});
-    }, {timeout: 120000, skipReason: ['Octave', 'NodeJS', 'Julia'].includes(lang) ? 'GROK-13876' : undefined});
+    }, {timeout: 120000});
 
     test('Dataframe performance test parallel', async () => {
       const iterations = DG.Test.isInBenchmark ? 10 : 3;
@@ -170,7 +173,7 @@ for (const lang of langs) {
       const sum = results.reduce((p, c) => p + c, 0);
       return toDart({'Average time': sum / results.length,
         'Min time': Math.min(...results), 'Max time': Math.max(...results)});
-    }, {timeout: 120000, skipReason: ['Octave', 'NodeJS', 'Julia'].includes(lang) ? 'GROK-13876' : undefined});
+    }, {timeout: 120000});
   });
 }