+++ packages/Charts/src/viewers/sunburst/sunburst-viewer.ts
@@ -21,6 +21,7 @@ export class SunburstViewer extends EChartViewer {
   hierarchyColumnNames: string[];
   hierarchyLevel: number;
   onClick: onClickOptions;
+  selectedOptions: string[] = ['Selected', 'SelectedOrCurrent', 'FilteredSelected'];
 
   constructor() {
     super();
@@ -65,11 +66,23 @@ export class SunburstViewer extends EChartViewer {
     }, event);
   }
 
-  createQueryMatcher(path: string[]): DG.RowMatcher {
-    const conditions = path.map((value, i) => {
-      return `${this.hierarchyColumnNames[i]} = ${value}`;
-    }).join(' and ');
-    return this.dataFrame.rows.match(conditions);
+  handleDataframeFiltering(path: string[], dataFrame: DG.DataFrame) {
+    const filterFunction = this.buildFilterFunction(path);
+    dataFrame.rows.filter(filterFunction);
+  }
+  
+  buildFilterFunction(path: string[]): (row: any) => boolean {
+    return (row) => {
+      for (let i = 0; i < path.length; ++i) {
+        const columnType = this.dataFrame.getCol(this.hierarchyColumnNames[i]).type;
+        const columnValue = row.get(this.hierarchyColumnNames[i]);
+        const formattedValue = columnType !== 'string' ? columnValue.toString() : columnValue;
+        const expectedValue = path[i];
+        if (formattedValue !== expectedValue)
+          return false;
+      }
+      return true;
+    };
   }
 
   removeFiltering() {
@@ -83,10 +96,10 @@ export class SunburstViewer extends EChartViewer {
       const selectedSectors: string[] = [];
       if (!params.data.path)
         return;
-      const path: string[] = params.data.path.split('|').map((str: string) => str.trim());
+      const path: string[] = params.treePathInfo.slice(1).map((obj: any) => obj.name);
       const pathString: string = path.join('|');
       if (this.onClick === 'Filter') {
-        this.createQueryMatcher(path).filter();
+        this.handleDataframeFiltering(path, this.dataFrame);
         return;
       }
       const isSectorSelected = selectedSectors.includes(pathString);
@@ -106,10 +119,14 @@ export class SunburstViewer extends EChartViewer {
         this.handleDataframeSelection(path, params.event.event);
       }
     });
-    this.chart.on('mouseover', (params: any) => {
-      const path: string[] = params.data.path.split('|').map((str: string) => str.trim());
-      const matchDf = this.createQueryMatcher(path).toDataFrame();
-      const matchCount = matchDf.rowCount;
+    this.chart.on('mouseover', async (params: any) => {
+      const path = params.treePathInfo.slice(1).map((obj: any) => obj.name);
+      const bitset = this.filter;
+      const matchDf = this.dataFrame.clone();
+      matchDf.rows.removeWhere(row => bitset && !bitset.get(row.idx));
+
+      this.handleDataframeFiltering(path, matchDf);
+      const matchCount = matchDf.filter.trueCount;
       ui.tooltip.showRowGroup(this.dataFrame, (i) => {
         const { hierarchyColumnNames, dataFrame } = this;
         for (let j = 0; j < hierarchyColumnNames.length; ++j) {
@@ -127,7 +144,13 @@ export class SunburstViewer extends EChartViewer {
         }
         return false;
       }, params.event.event.x, params.event.event.y);
-      ui.tooltip.root.innerText = `${matchCount}\n${params.name}`;
+      const {isSmiles, image} = await this.checkAndCreateMoleculeImage(params.name);
+      if (isSmiles && params.data.semType === 'Molecule') {
+        ui.tooltip.root.innerText = `${matchCount}`;
+        ui.tooltip.root.appendChild(image!);
+      } else {
+        ui.tooltip.root.innerText = `${matchCount}\n${params.name}`;  
+      }
     });      
     this.chart.on('mouseout', () => ui.tooltip.hide());
     this.chart.getDom().ondblclick = (event: MouseEvent) => {
@@ -165,14 +188,13 @@ export class SunburstViewer extends EChartViewer {
   onTableAttached(propertyChanged?: boolean): void {
     let categoricalColumns = [...this.dataFrame.columns.categorical].sort((col1, col2) =>
       col1.categories.length - col2.categories.length);
-    categoricalColumns = categoricalColumns.filter((col: DG.Column) => col.stats.missingValueCount != col.length);
+    categoricalColumns = categoricalColumns.filter((col: DG.Column) => col.stats.missingValueCount != col.length && !col.name.startsWith('~'));
 
     if (categoricalColumns.length < 1)
       return;
 
-    if (this.hierarchyColumnNames == null || this.hierarchyColumnNames.length === 0 || propertyChanged)
-      this.hierarchyColumnNames = categoricalColumns.slice(0, this.hierarchyLevel).map((col) => col.name);
-    
+    this.hierarchyColumnNames = categoricalColumns.slice(0, this.hierarchyLevel).map((col) => col.name);
+
     this.subs.push(this.dataFrame.onMetadataChanged.subscribe((_) => {this.render()}));
     this.subs.push(this.onContextMenu.subscribe(this.onContextMenuHandler.bind(this)));
     this.addSelectionOrDataSubs();
@@ -180,68 +202,86 @@ export class SunburstViewer extends EChartViewer {
   }
 
   getSeriesData(): treeDataType[] | undefined {
-    return TreeUtils.toForest(this.dataFrame, this.hierarchyColumnNames, this.filter);
+    const rowSource = this.selectedOptions.includes(this.rowSource!);
+    return TreeUtils.toForest(this.dataFrame, this.hierarchyColumnNames, this.filter, rowSource);
   }
 
   formatLabel(params: any) {
     //@ts-ignore
     const ItemAreaInfoArray = this.chart.getModel().getSeriesByIndex(0).getData()._itemLayouts.slice(1);
-    const getCurrentItemIndex = params.seriesIndex;
+    const getCurrentItemIndex = params.dataIndex - 1;
     const ItemLayoutInfo = ItemAreaInfoArray.find((item: any, index: number) => {
-        if (getCurrentItemIndex === index) {
-            return item;
-        }
+      if (getCurrentItemIndex === index)
+        return item;
     });
     const r = ItemLayoutInfo.r;
+    const r0 = ItemLayoutInfo.r0;
     const startAngle = ItemLayoutInfo.startAngle;
     const endAngle = ItemLayoutInfo.endAngle;
-    const cx = ItemLayoutInfo.cx;
-    const cy = ItemLayoutInfo.cy;
-    const width = Math.abs(cx + r * Math.cos(startAngle) - (cx + r * Math.cos(endAngle)));
-    const height = Math.abs(cy + r / 1.5 * Math.sin(startAngle) - (cy + r / 1.5 * Math.sin(endAngle)));
+    const {width, height} = this.calculateRingDimensions(r0, r, startAngle, endAngle);
 
-    const averageCharWidth = 10 * 0.6;
-    const averageCharHeight = 10 * 1.2;
+    const averageCharWidth = 5;
+    const averageCharHeight = 10;
     const maxWidthCharacters = Math.floor(width / averageCharWidth);
     const maxHeightCharacters = Math.floor(height / averageCharHeight);
-
     const maxLength = maxWidthCharacters;
-    let name = params.name;
-    let lines = [name];
 
-    if (name.length > maxLength) {
-      lines = [];
-      let remainingHeight = maxHeightCharacters;
-      while (name.length > 0 && remainingHeight > 0) {
-        let line = name.substring(0, maxLength);
-        const lastSpaceIndex = line.lastIndexOf(' ');
-        if (lastSpaceIndex !== -1) {
-          line = line.substring(0, lastSpaceIndex);
-        }
-        line = line.trimRight();
-        lines.push(line);  
-        remainingHeight -= 1;
-        name = name.substring(line.length).trim();
+    const name = params.name;
+    const lines = name.split(' ');
+    let result = '';
+    let remainingHeight = maxHeightCharacters;
+
+    for (let line of lines) {
+      if (line.length > maxLength || remainingHeight <= 0) {
+        result = '';
+        break;
       }
-      if (name.length > 0) {
-        lines[lines.length - 1] += '...';
+      if (result.length > 0) {
+        result += '\n';
+        remainingHeight--;
       }
+      result += line;
+      if (result.length >= maxLength || remainingHeight <= 0)
+        break;
     }
-    return lines.join('\n');
+
+    if (result.length < name.length)
+      result += '...';
+
+    const resultWidth = result.length * averageCharWidth;
+    const resultHeight = result.split('\n').length * averageCharHeight;
+    if (resultWidth > width || resultHeight > height)
+      result = '...';
+    
+    return result;
+  }
+
+  calculateRingDimensions(innerRadius: number, outerRadius: number, startAngle: number, endAngle: number) {
+    let width = outerRadius - innerRadius;
+    let height = Math.abs(endAngle - startAngle) * outerRadius;
+    return { height, width };
+  }
+
+  async checkAndCreateMoleculeImage(name: string): Promise<{ isSmiles: boolean, image: HTMLCanvasElement | null }> {
+    const isSmiles = await grok.functions.call('Chem:isSmiles', {s: name});
+    let image: HTMLCanvasElement | null = null;
+    if (isSmiles) {
+      const imageContainer = await grok.functions.call('Chem:drawMolecule', {
+        'molStr': name, 'w': 70, 'h': 80, 'popupMenu': false
+      });
+      image = imageContainer.querySelector(".chem-canvas");
+    }
+    return {isSmiles, image};
   }
 
   async handleStructures(data: treeDataType[] | undefined) {
     for (const entry of data!) {
       const name = entry.name;
-      const isSmiles = await grok.functions.call('Chem:isSmiles', {s: name});
+      const { isSmiles, image } = await this.checkAndCreateMoleculeImage(name);
       if (isSmiles && entry.semType === 'Molecule') {
-        const imageContainer = await grok.functions.call('Chem:drawMolecule', {
-          'molStr': name, 'w': 70, 'h': 80, 'popupMenu': false
-        });
-        const image = imageContainer.querySelector(".chem-canvas");
         await delay(5);
         const img = new Image();
-        img.src = image.toDataURL('image/png');
+        img.src = image!.toDataURL('image/png');
         entry.label = {
           show: true,
           formatter: '{b}',