+++ packages/Charts/src/viewers/tree/tree-viewer.ts
@@ -129,7 +129,7 @@ export class TreeViewer extends EChartViewer {
     }
     if (p?.name === 'table') {
       this.updateTable();
-      this.onTableAttached(true);
+      this.onTableAttached();
       this.render();
     }
     if (p?.name === 'hierarchyColumnNames' || p?.name === 'sizeColumnName' ||
@@ -152,15 +152,17 @@ export class TreeViewer extends EChartViewer {
     return isColumnNumerical || (!isColumnNumerical && isAggregationApplicable);
   }
   
-  onTableAttached(propertyChanged?: boolean) {
+  onTableAttached() {
     const categoricalColumns = [...this.dataFrame.columns.categorical].sort((col1, col2) =>
       col1.categories.length - col2.categories.length);
 
     if (categoricalColumns.length < 1)
       return;
 
-    if (this.hierarchyColumnNames == null || this.hierarchyColumnNames.length === 0 || propertyChanged)
-      this.hierarchyColumnNames = categoricalColumns.slice(0, 3).map((col) => col.name);
+    this.filter = this.dataFrame.filter;
+    this.hierarchyColumnNames = categoricalColumns.slice(0, 3).map((col) => col.name);
+    this.sizeColumnName = '';
+    this.colorColumnName = '';
 
     super.onTableAttached();
     this.initChartEventListeners();