+++ packages/Compute/scripts/object-cooling.js
@@ -20,6 +20,7 @@
 //meta.foldedCategories: ["Object properties"]
 //meta.compareCustomizer: Compute:CustomCustomizer
 //meta.uploadFunc: Compute:CustomUploader
+//meta.help: Compute/readme.md
 
 timeToCool = undefined;