+++ packages/DBTests/queries/mariadb-pattern-test.sql
@@ -14,21 +14,21 @@ SELECT * FROM mock_data WHERE id = @id;
 -- name: MariaDbStringTypeIntPatternOpMore
 -- connection: MariaDbDBTests
 -- input: string id = ">28" {pattern: int}
--- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpMore(), OpenFile('System:AppData/Dbtests/common/data29-30.d42')) // cat: MariaDb
+-- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpMore('>28'), OpenFile('System:AppData/Dbtests/common/data29-30.d42')) // cat: MariaDb
 SELECT * FROM mock_data WHERE @id(id)
 -- end
 
 -- name: MariaDbStringTypeIntPatternOpMoreEq
 -- connection: MariaDbDBTests
 -- input: string id = ">=29" {pattern: int}
--- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpMoreEq(), OpenFile('System:AppData/Dbtests/common/data29-30.d42')) // cat: MariaDb
+-- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpMoreEq('>=29'), OpenFile('System:AppData/Dbtests/common/data29-30.d42')) // cat: MariaDb
 SELECT * FROM mock_data WHERE @id(id)
 -- end
 
 -- name: MariaDbStringTypeIntPatternOpLessEq
 -- connection: MariaDbDBTests
 -- input: string id = "<=1" {pattern: int}
--- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpLessEq(), OpenFile('System:AppData/Dbtests/common/data1.d42')) // cat: MariaDb
+-- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpLessEq('<=1'), OpenFile('System:AppData/Dbtests/common/data1.d42')) // cat: MariaDb
 SELECT * FROM mock_data WHERE @id(id)
 --
 -- end
@@ -36,35 +36,35 @@ SELECT * FROM mock_data WHERE @id(id)
 -- name: MariaDbStringTypeIntPatternOpLess
 -- connection: MariaDbDBTests
 -- input: string id = "<2" {pattern: int}
--- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpLess(), OpenFile('System:AppData/Dbtests/common/data1.d42')) // cat: MariaDb
+-- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpLess('<2'), OpenFile('System:AppData/Dbtests/common/data1.d42')) // cat: MariaDb
 SELECT * FROM mock_data WHERE @id(id)
 -- end
 
 -- name: MariaDbStringTypeIntPatternOpIn
 -- connection: MariaDbDBTests
 -- input: string id = "in(29, 30)" {pattern: int}
--- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpIn(), OpenFile('System:AppData/Dbtests/common/data29-30.d42')) // cat: MariaDb
+-- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpIn(id='in(29, 30)'), OpenFile('System:AppData/Dbtests/common/data29-30.d42')) // cat: MariaDb
 SELECT * FROM mock_data WHERE @id(id)
 -- end
 
 -- name: MariaDbStringTypeIntPatternOpNotIn
 -- connection: MariaDbDBTests
 -- input: string id = "not in(21, 22, 23, 24, 25, 26, 27, 28, 29, 30)" {pattern: int}
--- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpNotIn(), OpenFile('System:AppData/Dbtests/common/data1-20.d42')) // cat: MariaDb
+-- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpNotIn(id='not in(21, 22, 23, 24, 25, 26, 27, 28, 29, 30)'), OpenFile('System:AppData/Dbtests/common/data1-20.d42')) // cat: MariaDb
 SELECT * FROM mock_data WHERE @id(id)
 -- end
 
 -- name: MariaDbStringTypeIntPatternOpMinMax
 -- connection: MariaDbDBTests
 -- input: string id = "min-max 29-30" {pattern: int}
--- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpMinMax(), OpenFile('System:AppData/Dbtests/common/data29-30.d42')) // cat: MariaDb
+-- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpMinMax(id='min-max 29-30'), OpenFile('System:AppData/Dbtests/common/data29-30.d42')) // cat: MariaDb
 SELECT * FROM mock_data WHERE @id(id)
 -- end
 
 -- name: MariaDbStringTypeIntPatternOpNotEq
 -- connection: MariaDbDBTests
 -- input: string id = "!=1" {pattern: int}
--- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpNotEq(), OpenFile('System:AppData/Dbtests/common/data2-30.d42')) // cat: MariaDb
+-- test: Dbtests:expectTable(MariaDbStringTypeIntPatternOpNotEq('!=1'), OpenFile('System:AppData/Dbtests/common/data2-30.d42')) // cat: MariaDb
 SELECT * FROM mock_data WHERE @id(id)
 -- end
 
@@ -78,28 +78,28 @@ SELECT * FROM mock_data WHERE some_number = @some_number;
 -- name: MariaDbStringTypePatternDoubleOpMore
 -- connection: MariaDbDBTests
 -- input: string some_number = ">975" {pattern: double}
--- test: Dbtests:expectTable(MariaDbStringTypePatternDoubleOpMore(), OpenFile('System:AppData/Dbtests/common/data10,26.d42')) // cat: MariaDb
+-- test: Dbtests:expectTable(MariaDbStringTypePatternDoubleOpMore('>975'), OpenFile('System:AppData/Dbtests/common/data10,26.d42')) // cat: MariaDb
 SELECT * FROM mock_data WHERE @some_number(some_number);
 -- end
 
 -- name: MariaDbStringTypePatternDoubleOpMoreEq
 -- connection: MariaDbDBTests
 -- input: string some_number = ">=975" {pattern: double}
--- test: Dbtests:expectTable(MariaDbStringTypePatternDoubleOpMoreEq(), OpenFile('System:AppData/Dbtests/common/data10,26.d42')) // cat: MariaDb
+-- test: Dbtests:expectTable(MariaDbStringTypePatternDoubleOpMoreEq('>=975'), OpenFile('System:AppData/Dbtests/common/data10,26.d42')) // cat: MariaDb
 SELECT * FROM mock_data WHERE @some_number(some_number);
 -- end
 
 -- name: MariaDbStringTypePatternDoubleOpLess
 -- connection: MariaDbDBTests
 -- input: string some_number = "<20" {pattern: double}
--- test: Dbtests:expectTable(MariaDbStringTypePatternDoubleOpLess(), OpenFile('System:AppData/Dbtests/common/data5.d42')) // cat: MariaDb
+-- test: Dbtests:expectTable(MariaDbStringTypePatternDoubleOpLess('<20'), OpenFile('System:AppData/Dbtests/common/data5.d42')) // cat: MariaDb
 SELECT * FROM mock_data WHERE @some_number(some_number);
 -- end
 
 -- name: MariaDbStringTypePatternDoubleOpLessEq
 -- connection: MariaDbDBTests
 -- input: string some_number = "<=20" {pattern: double}
--- test: Dbtests:expectTable(MariaDbStringTypePatternDoubleOpLessEq(), OpenFile('System:AppData/Dbtests/common/data5.d42')) // cat: MariaDb
+-- test: Dbtests:expectTable(MariaDbStringTypePatternDoubleOpLessEq('<=20'), OpenFile('System:AppData/Dbtests/common/data5.d42')) // cat: MariaDb
 SELECT * FROM mock_data WHERE @some_number(some_number);
 -- end
 
@@ -127,7 +127,7 @@ SELECT * FROM mock_data WHERE @first_name(first_name);
 -- name: MariaDbStringTypePatternStringOpIn
 -- connection: MariaDbDBTests
 -- input: string country = "in (Poland, Brazil)" {pattern: string}
--- test: Dbtests:expectTable(MariaDbStringTypePatternStringOpIn(), OpenFile('System:AppData/Dbtests/common/data2,5,20.d42')) // cat: MariaDb
+-- test: Dbtests:expectTable(MariaDbStringTypePatternStringOpIn(country='in (Poland, Brazil)'), OpenFile('System:AppData/Dbtests/common/data2,5,20.d42')) // cat: MariaDb
 SELECT * FROM mock_data WHERE @country(country);
 -- end