+++ packages/Dendrogram/src/utils/tree-helper.ts
@@ -5,14 +5,15 @@ import * as DG from 'datagrok-api/dg';
 import wu from 'wu';
 import {DistanceMetric, isLeaf, NodeCuttedType, NodeType} from '@datagrok-libraries/bio/src/trees';
 import {NO_NAME_ROOT, parseNewick} from '@datagrok-libraries/bio/src/trees/phylocanvas';
+import {NEWICK_EMPTY} from '@datagrok-libraries/bio/src/trees/consts';
 import {DistanceMatrix, DistanceMatrixService} from '@datagrok-libraries/ml/src/distance-matrix';
 import {ITreeHelper} from '@datagrok-libraries/bio/src/trees/tree-helper';
 import {ClusterMatrix} from '@datagrok-libraries/bio/src/trees';
 import {UnitsHandler} from '@datagrok-libraries/bio/src/utils/units-handler';
-import {MmDistanceFunctionsNames} from
-  '@datagrok-libraries/ml/src/macromolecule-distance-functions';
+import {MmDistanceFunctionsNames} from '@datagrok-libraries/ml/src/macromolecule-distance-functions';
 import {NumberMetricsNames} from '@datagrok-libraries/ml/src/typed-metrics';
 import {IntArrayMetricsNames} from '@datagrok-libraries/ml/src/typed-metrics/consts';
+
 type DataNodeDict = { [nodeName: string]: number };
 
 export const enum TAGS {
@@ -52,17 +53,17 @@ export class TreeHelper implements ITreeHelper {
         name = `${nodePrefixV}${NO_NAME_ROOT}`;
       }
 
+      if (!!obj.children) {
+        const childrenCount = obj.children.length;
+        for (let i = 0; i < childrenCount; i++)
+          traverse(obj.children[i], obj);
+      }
+
       nodes.push(name);
       distances.push(obj.branch_length ? obj.branch_length : null);
       // annotations.push(obj.annotation);
       parents.push(parent ? parent!.name : null);
       leafs.push(!obj.children || obj.children.length == 0);
-
-      if (!obj.children) return;
-      const childrenNum = obj.children.length;
-
-      for (let i = 0; i < childrenNum; i++)
-        traverse(obj.children[i], obj);
     }
 
     traverse(obj, null);
@@ -111,7 +112,7 @@ export class TreeHelper implements ITreeHelper {
       }
     }
 
-    return !node ? ';' : `${toNewickInt(node)};`;
+    return !node ? NEWICK_EMPTY : `${toNewickInt(node)};`;
   }
 
   getLeafList<TNode extends NodeType>(node: TNode | null, list?: TNode[]): TNode[] {
@@ -267,7 +268,7 @@ export class TreeHelper implements ITreeHelper {
    * @param {string}leafColName - name of column with leaf names
    * @param {boolean}removeMissingDataRows - remove rows with missing data
    * @return {[NodeType, string[]]} - tree (root node) of nodes presented in data and list of missed data nodes
-  */
+   */
   setGridOrder(
     tree: NodeType | null, grid: DG.Grid, leafColName?: string,
     removeMissingDataRows: boolean = false,
@@ -555,8 +556,10 @@ export class TreeHelper implements ITreeHelper {
         let newMat: DistanceMatrix | null = new DistanceMatrix(values);
         newMat.normalize();
         switch (method) {
-        case DistanceMetric.Manhattan:
+        case DistanceMetric.Manhattan: {
           out.add(newMat);
+          break;
+        }
         default:
           newMat.square();
           out.add(newMat);
@@ -571,6 +574,7 @@ export class TreeHelper implements ITreeHelper {
 
     return out;
   }
+
   parseClusterMatrix(clusterMatrix: ClusterMatrix): NodeType {
     /*
     clusert matrix is in R format, I.E. the indexings are 1-based.
@@ -586,6 +590,7 @@ export class TreeHelper implements ITreeHelper {
       function subTreeLength(children?: NodeType[]): number {
         return children && children.length ? (children[0].branch_length ?? 0) + subTreeLength(children[0].children) : 0;
       }
+
       if (isLeaf(node))
         return 0;
       else