+++ packages/PowerGrid/src/sparklines/piechart.ts
@@ -10,7 +10,7 @@ import {
   isSummarySettingsBase
 } from './shared';
 
-const minRadius = 10;
+let minRadius: number;
 
 enum PieChartStyle {
   Radius = 'Radius',
@@ -20,12 +20,18 @@ enum PieChartStyle {
 interface PieChartSettings extends SummarySettingsBase {
   radius: number;
   style: PieChartStyle.Radius | PieChartStyle.Angle;
+  sectors: {
+    sectorColor: string;
+    subsectors: { name: string; radius: number }[];
+  }[];
 }
 
 function getSettings(gc: DG.GridColumn): PieChartSettings {
+  const sectors = gc.settings.sectors;
   const settings: PieChartSettings = isSummarySettingsBase(gc.settings) ? gc.settings :
     gc.settings[SparklineType.PieChart] ??= getSettingsBase(gc, SparklineType.PieChart);
   settings.style ??= PieChartStyle.Radius;
+  settings.sectors ??= sectors;
   return settings;
 }
 
@@ -50,11 +56,37 @@ function onHit(gridCell: DG.GridCell, e: MouseEvent): Hit {
   let activeColumn = -1;
   const row: number = gridCell.cell.row.idx;
 
-  let r: number;
-  if (settings.style == PieChartStyle.Radius) {
+  let r: number = (gridCell.bounds.width - 4) / 2;
+  if (settings.style == PieChartStyle.Radius && !settings.sectors) {
     activeColumn = Math.floor((angle * cols.length) / (2 * Math.PI));
     r = cols[activeColumn].scale(row) * (gridCell.bounds.width - 4) / 2;
-    r = r < minRadius ? minRadius : r;
+    r = Math.max(r, minRadius);
+  } else if (settings.sectors) {
+    const sectors = settings.sectors;
+    const totalSectors = sectors.length;
+    const sectorAngle = (2 * Math.PI) / totalSectors;
+    const normalizedAngle = (angle + Math.PI * 2) % (Math.PI * 2);
+    for (let i = 0; i < totalSectors; ++i) {
+      const sector = sectors[i];
+      const sectorStartAngle = i * sectorAngle;
+      const sectorEndAngle = (i + 1) * sectorAngle;
+      if (normalizedAngle >= sectorStartAngle && normalizedAngle < sectorEndAngle) {
+        const subsectors = sector.subsectors;
+        const totalSubsectors = subsectors.length;
+        const subsectorAngle = sectorAngle / totalSubsectors;
+        let subsectorStartAngle = sectorStartAngle;
+        for (let j = 0; j < totalSubsectors; ++j) {
+          const subsector = subsectors[j];
+          const subsectorEndAngle = subsectorStartAngle + subsectorAngle;
+          if (normalizedAngle >= subsectorStartAngle && normalizedAngle < subsectorEndAngle) {
+            activeColumn = cols.findIndex((col) => col && subsector && col.name === subsector.name);
+            break;
+          }
+          subsectorStartAngle = subsectorEndAngle;
+        }
+        break;
+      }
+    }
   } else {
     const sum = getColumnsSum(cols, row);
     r = (gridCell.bounds.width - 4) / 2;
@@ -115,13 +147,14 @@ export class PieChartCellRenderer extends DG.GridCellRenderer {
     const row: number = gridCell.cell.row.idx;
     const cols = df.columns.byNames(settings.columnNames);
     const box = new DG.Rect(x, y, w, h).fitSquare().inflate(-2, -2);
-    if (settings.style == PieChartStyle.Radius) {
+    minRadius = Math.min(box.width, box.height) / 10;
+    if (settings.style == PieChartStyle.Radius && !settings.sectors) {
       for (let i = 0; i < cols.length; i++) {
         if (cols[i].isNone(row))
           continue;
 
         let r = cols[i].scale(row) * box.width / 2;
-        r = r < minRadius ? minRadius : r;
+        r = Math.max(r, minRadius);
         g.beginPath();
         g.moveTo(box.midX, box.midY);
         g.arc(box.midX, box.midY, r,
@@ -132,6 +165,41 @@ export class PieChartCellRenderer extends DG.GridCellRenderer {
         g.fill();
         g.strokeStyle = DG.Color.toRgb(DG.Color.lightGray);
         g.stroke();
+      }
+    } else if (settings.sectors) {
+      let currentAngle = 0;
+      const sectors = settings.sectors;
+      for (let i = 0; i < sectors.length; ++i) {
+        const sector = sectors[i];
+        const sectorAngle = (2 * Math.PI) / sectors.length;
+        g.beginPath();
+        g.moveTo(box.midX, box.midY);
+        g.arc(box.midX, box.midY, Math.min(box.width, box.height) / 2, currentAngle, currentAngle + sectorAngle);
+        g.closePath();
+        g.fillStyle = this.hexToRgbA(sector.sectorColor, 0.4);
+        g.fill();
+        
+        const totalSubsectors = sector.subsectors.length;
+        let subsectorCurrentAngle = currentAngle;
+        sector.subsectors.forEach(subsector => {
+          const subsectorAngle = sectorAngle / totalSubsectors;
+          const gap = 0.05;
+          let r = Math.max(Math.min(box.width, box.height) / 2, minRadius);
+          const subsectorName = subsector.name;
+          const subsectorCol = cols.find(col => col.name === subsectorName);
+          if (subsectorCol) {
+            r = subsectorCol.scale(row) * (Math.min(box.width, box.height) / 2);
+            r = Math.max(r, minRadius);
+          }
+          g.beginPath();
+          g.moveTo(box.midX, box.midY);
+          g.arc(box.midX, box.midY, r, subsectorCurrentAngle + gap, subsectorCurrentAngle + subsectorAngle - gap);
+          g.closePath();
+          g.fillStyle = this.hexToRgbA(sector.sectorColor, 0.8);
+          g.fill();
+          subsectorCurrentAngle += subsectorAngle;
+        });
+        currentAngle += sectorAngle;
       }
     } else {
       const sum = getColumnsSum(cols, row);
@@ -156,6 +224,14 @@ export class PieChartCellRenderer extends DG.GridCellRenderer {
     }
   }
 
+  hexToRgbA(hex: string, opacity: number): string {
+    const bigint = parseInt(hex.substring(1), 16);
+    const r = (bigint >> 16) & 255;
+    const g = (bigint >> 8) & 255;
+    const b = bigint & 255;
+    return `rgba(${r},${g},${b},${opacity})`;
+  }
+  
   renderSettings(gc: DG.GridColumn): Element {
     const settings: PieChartSettings = isSummarySettingsBase(gc.settings) ? gc.settings :
       gc.settings[SparklineType.PieChart] ??= getSettings(gc);